Unify testing core and plugins
